20170709{待完善}

+++ app/src/main/java/com/zucc/zwy1317/myassistant/ui/customview/agenda/AgendaHeaderView.java
@@ -85,10 +85,10 @@ public class AgendaHeaderView extends LinearLayout {
 
         //得到每个月的第一天
         List<Calendar> list = CalendarManager.getInstance().getFirstDayOfMonth();
-        if (day.get(Calendar.DAY_OF_MONTH) == 1) {
+        if ((int)day.get(Calendar.DAY_OF_MONTH) == 1) {
             System.out.println("---@@--" + day.get(Calendar.DAY_OF_MONTH) + "--" + day.get(Calendar.MONTH));
             for (int i = 0; i < list.size(); i++) {
-                if (day.get(Calendar.MONTH) == i) {
+                if ((int)day.get(Calendar.MONTH) == i) {
                     System.out.println("----iiii---" + dip2px(120));
                     setLayoutParams(new FrameLayout.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, dip2px(120)));
